% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calculate_energy}
\alias{calculate_energy}
\title{Global energy of the model}
\usage{
calculate_energy(
  idx,
  seg,
  pop,
  walk,
  bicy,
  rts_walk = NULL,
  rts_bicy = NULL,
  od = NULL
)
}
\arguments{
\item{idx}{numeric, indices of candidates ins 'seg'.}

\item{seg}{sf, road segments point locations.}

\item{pop}{sf, population point data.}

\item{walk}{dodgr routing graph, graph to route walking times between the points.}

\item{bicy}{dodgr routing graph, graph to route bicycle times between the points.}

\item{rts_walk, }{data.table, precalculated walking to station routes; OD matrix (default = NULL).}

\item{rts_bicy, }{data.table, precalculated station to station bicycle routes; OD matrix (default = NULL).}

\item{od, }{data.table, precalculated direct walking routes; OD matrix (default = NULL).}
}
\value{
A numeric scalar with the global energy.
}
\description{
Energy function that calculates the global energy of the model. The energy
quantifies the duration of all origin destination travels by bicycle (walk
to station, cycle, walk from station) or alternatively the direct walking
travel if it is faster. The travels are weighted by the origin and
destination population.
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

# Get variables
idx <- m$idx
seg <- m$layer$seg
pop <- m$layer$pop
walk <- m$route$walk
bicy <- m$route$bicy

# Global energy
calculate_energy(idx, seg, pop, walk, bicy)
}
