% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_summary}
\alias{drt_summary}
\title{Print a summary of the model performance}
\usage{
drt_summary(obj, walking_limit = 10)
}
\arguments{
\item{obj, }{drtm, a drtm model.}

\item{walking_limit, }{numeric, walking time limit in minutes to split the summary.}
}
\value{
None.
}
\description{
Print a summary of the model performance
}
\examples{
# Example model
m <- drt_import(
  system.file("example_i1000.RData", package = "drtplanr")
)

drt_summary(m)
}
