% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.sbm}
\alias{print.sbm}
\title{Print}
\usage{
\method{print}{sbm}(x, ...)
}
\arguments{
\item{x}{sbplanr object, print information about package classes.}

\item{...}{...}
}
\value{
None.
}
\description{
Print
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)
print(m)
}
