% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_energy<-}
\alias{sb_energy<-}
\title{Set energy function}
\usage{
sb_energy(x) <- value
}
\arguments{
\item{x, }{sbm, a sbm model.}

\item{value}{function, energy function to calculate energy of the model.}
}
\value{
A new sbm with the new energy.
}
\description{
Set energy function
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

sb_energy(m) <- calculate_energy
}
