% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_energy}
\alias{sb_energy}
\title{Get energy function}
\usage{
sb_energy(x)
}
\arguments{
\item{x, }{sbm, a sbm model.}
}
\value{
The energy function of the model.
}
\description{
Get energy function
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

sb_energy(m)
}
