% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_export}
\alias{sb_export}
\title{Export sbm}
\usage{
sb_export(obj, path)
}
\arguments{
\item{obj}{sbm, a sbm model of the sbplanr.}

\item{path}{character, path to file.}
}
\value{
None.
}
\description{
Export sbm
}
\examples{
\donttest{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

# Export to temporary dir
sb_export(m, path = tempdir())
}
}
