% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_import}
\alias{sb_import}
\title{Import sbm}
\usage{
sb_import(file_name)
}
\arguments{
\item{file_name}{character, path to file.}
}
\value{
A sbm model.
}
\description{
Import sbm
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)
}
