% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_iterate}
\alias{sb_iterate}
\title{Iterate}
\usage{
sb_iterate(obj, n_iter, precalculate = TRUE, annealing = TRUE)
}
\arguments{
\item{obj, }{sbm, a sbm model.}

\item{n_iter}{numeric, number of iterations.}

\item{precalculate, }{boolean, precalculate walking and cycling times? Speeds up iterations, but needs more memory (defaul = TRUE).}

\item{annealing}{boolean, apply annealing (alpha = 1/(i+1)) (default = TRUE).}
}
\value{
A new sbm with 'n_iter' times more iterations.
}
\description{
Minimize the gloabal energy of a sbm model.
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

sb_iterate(m, 3)
}
