% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_map}
\alias{sb_map}
\title{Plot station map}
\usage{
sb_map(obj)
}
\arguments{
\item{obj}{sbm, a sbm model.}
}
\value{
A mapview map object, containing the energy curve.
}
\description{
Plot station map
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

# sb_map(m)
}
