% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_osm_bb}
\alias{sb_osm_bb}
\title{Get the box in the osmdata format.}
\usage{
sb_osm_bb(aoi)
}
\arguments{
\item{aoi, }{spatial object, object to retreive the bounding box from.}
}
\value{
A matrix conaining the bbox in osmdata format.
}
\description{
Get the box in the osmdata format.
}
\examples{
# Example data
aoi <-
  sf::st_read(system.file("example.gpkg", package = "sbplanr"), layer = "aoi")

sb_osm_bb(aoi)
}
