% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_plot}
\alias{sb_plot}
\title{Plot energy curve}
\usage{
sb_plot(obj)
}
\arguments{
\item{obj}{sbm, a sbm model.}
}
\value{
A ggplot2 plor object, containing the energy curve.
}
\description{
Plot energy curve
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

# Plot
sb_plot(m)
}
