% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_reset}
\alias{sb_reset}
\title{Reset the model state}
\usage{
sb_reset(obj, n_sta = NULL, shuffle = FALSE)
}
\arguments{
\item{obj, }{sbm, a sbm model.}

\item{n_sta, }{numeric, number of stations (default = NULL).}

\item{shuffle, }{boolean, shuffle the initial station positions?}
}
\value{
The energy function of the model.
}
\description{
Reset the model state
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

sb_reset(m)
}
