% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_roa_seg}
\alias{sb_roa_seg}
\title{Segmentize roads}
\usage{
sb_roa_seg(roa, m_seg)
}
\arguments{
\item{roa}{spatial object, contains the roads.}

\item{m_seg}{numeric, resoution for the length of the segemnts in meters.}
}
\value{
A sf object containing the segementized road network points.
}
\description{
Segmentize roads
}
\examples{
print("tbd.")
}
