% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sb_summary}
\alias{sb_summary}
\title{Print a summary of the model performance}
\usage{
sb_summary(obj, walking_limit = 5)
}
\arguments{
\item{obj, }{sbm, a sbm model.}

\item{walking_limit, }{numeric, walking time limit in minutes to split the summary (default = 5).}
}
\value{
None.
}
\description{
Print a summary of the model performance
}
\examples{
# Example model
m <- sb_import(
  system.file("example_i1000.RData", package = "sbplanr")
)

sb_summary(m)
}
