% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbplanr-package.R
\docType{package}
\name{sbplanr-package}
\alias{sbplanr}
\alias{sbplanr-package}
\title{sbplanr: Experimental station-based bicycle sharing planner}
\description{
Tool for placing bicycle sharing stations by iteratively minimizing
    a global energy of a model that reflects the station-based bicycle sharing
    system. The station locations are randomly initialized in the street network
    and iteratively optimized based on the reachable population in combination
    with walking and driving times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munterfi/sbplanr/}
  \item Report bugs at \url{https://github.com/munterfi/sbplanr/issues/}
}

}
\author{
\strong{Maintainer}: Merlin Unterfinger \email{info@munterfinger.ch} (\href{https://orcid.org/0000-0003-2020-2366}{ORCID})

}
\keyword{internal}
